+++ tf.go
@@ -2,7 +2,6 @@ package goyek
 
 import (
 	"context"
-	"fmt"
 	"io"
 	"os"
 	"os/exec"
@@ -18,12 +17,12 @@ import (
 // All methods must be called only from the goroutine running the
 // Action function.
 type TF struct {
-	ctx       context.Context
-	name      string
-	output    io.Writer
-	decorator LogDecorator
-	failed    bool
-	skipped   bool
+	ctx     context.Context
+	name    string
+	output  io.Writer
+	logger  Logger
+	failed  bool
+	skipped bool
 }
 
 // Context returns the flows' run context.
@@ -56,25 +55,25 @@ func (tf *TF) Cmd(name string, args ...string) *exec.Cmd {
 // and prints the text to Output. A final newline is added.
 // The text will be printed only if the task fails or flow is run in Verbose mode.
 func (tf *TF) Log(args ...interface{}) {
-	tf.log(args...)
+	tf.logger.Log(tf.output, args...)
 }
 
 // Logf formats its arguments according to the format, analogous to Printf,
 // and prints the text to Output. A final newline is added.
 // The text will be printed only if the task fails or flow is run in Verbose mode.
 func (tf *TF) Logf(format string, args ...interface{}) {
-	tf.logf(format, args...)
+	tf.logger.Logf(tf.output, format, args...)
 }
 
 // Error is equivalent to Log followed by Fail.
 func (tf *TF) Error(args ...interface{}) {
-	tf.log(args...)
+	tf.logger.Log(tf.output, args...)
 	tf.Fail()
 }
 
 // Errorf is equivalent to Logf followed by Fail.
 func (tf *TF) Errorf(format string, args ...interface{}) {
-	tf.logf(format, args...)
+	tf.logger.Logf(tf.output, format, args...)
 	tf.Fail()
 }
 
@@ -90,13 +89,13 @@ func (tf *TF) Fail() {
 
 // Fatal is equivalent to Log followed by FailNow.
 func (tf *TF) Fatal(args ...interface{}) {
-	tf.log(args...)
+	tf.logger.Log(tf.output, args...)
 	tf.FailNow()
 }
 
 // Fatalf is equivalent to Logf followed by FailNow.
 func (tf *TF) Fatalf(format string, args ...interface{}) {
-	tf.logf(format, args...)
+	tf.logger.Logf(tf.output, format, args...)
 	tf.FailNow()
 }
 
@@ -116,13 +115,13 @@ func (tf *TF) Skipped() bool {
 
 // Skip is equivalent to Log followed by SkipNow.
 func (tf *TF) Skip(args ...interface{}) {
-	tf.log(args...)
+	tf.logger.Log(tf.output, args...)
 	tf.SkipNow()
 }
 
 // Skipf is equivalent to Logf followed by SkipNow.
 func (tf *TF) Skipf(format string, args ...interface{}) {
-	tf.logf(format, args...)
+	tf.logger.Logf(tf.output, format, args...)
 	tf.SkipNow()
 }
 
@@ -163,18 +162,4 @@ func (tf *TF) run(action func(tf *TF)) Result {
 		finished = true
 	}()
 	return <-ch
-}
-
-// log is used internally in order to provide proper prefix.
-func (tf *TF) log(args ...interface{}) {
-	txt := fmt.Sprint(args...)
-	txt = tf.decorator.Decorate(txt)
-	io.WriteString(tf.output, txt) //nolint:errcheck,gosec // not checking errors when writing to output
-}
-
-// lof is used internally in order to provide proper prefix.
-func (tf *TF) logf(format string, args ...interface{}) {
-	txt := fmt.Sprintf(format, args...)
-	txt = tf.decorator.Decorate(txt)
-	io.WriteString(tf.output, txt) //nolint:errcheck,gosec // not checking errors when writing to output
 }